+++ src/pczt/prover.rs
@@ -57,7 +57,7 @@ impl super::Bundle {
                 let spend =
                     SpendInfo::new(fvk, note, merkle_path).ok_or(ProverError::WrongFvkForNote)?;
 
-                let output_note = Note::from_parts(
+                let output_note = Note::from_parts_with_version(
                     action
                         .output
                         .recipient
@@ -65,6 +65,7 @@ impl super::Bundle {
                     action.output.value.ok_or(ProverError::MissingValue)?,
                     Rho::from_nf_old(action.spend.nullifier),
                     action.output.rseed.ok_or(ProverError::MissingRandomSeed)?,
+                    action.output.note_version,
                 )
                 .into_option()
                 .ok_or(ProverError::InvalidOutputNote)?;